+++ eNMS/server.py
@@ -1,4 +1,4 @@
-from click import argument, echo, option
+from click import argument, echo, option, Choice
 from datetime import datetime, timedelta
 from json import loads
 from passlib.hash import argon2
@@ -19,6 +19,7 @@ from flask_login import current_user, LoginManager, login_user, logout_user
 from flask_restful import abort as rest_abort, Api, Resource
 from flask_wtf.csrf import CSRFProtect
 from functools import wraps
+from getpass import getuser
 from itertools import chain
 from logging import info
 from os import environ
@@ -320,25 +321,30 @@ class Server(Flask):
             devices_list = devices.split(",") if devices else []
             devices_list = [db.fetch("device", name=name).id for name in devices_list]
             payload_dict = loads(payload) if payload else {}
-            payload_dict["devices"] = devices_list
+            payload_dict.update(devices=devices_list, trigger="CLI", creator=getuser())
             service = db.fetch("service", name=name)
             results = app.run(service.id, **payload_dict)
             db.session.commit()
             echo(app.str_dict(results))
 
-        @self.cli.command(name="delete-changelog")
+        @self.cli.command(name="delete_log")
         @option(
-            "--keep-last-days",
-            default=15,
-            help="Number of days to keep in the changelog",
+            "--keep-last-days", default=15, help="Number of days to keep",
         )
-        def remove_changelog(keep_last_days):
+        @option(
+            "--log",
+            "-l",
+            required=True,
+            type=Choice(("changelog", "result")),
+            help="Type of logs",
+        )
+        def delete_log(keep_last_days, log):
             deletion_time = datetime.now() - timedelta(days=keep_last_days)
             app.result_log_deletion(
                 date_time=deletion_time.strftime("%d/%m/%Y %H:%M:%S"),
-                deletion_types=["changelog"],
+                deletion_types=[log],
             )
-            app.log("info", f"deleted all changelogs up until {deletion_time}")
+            app.log("info", f"deleted all logs in '{log}' up until {deletion_time}")
 
     def configure_rest_api(self):