+++ docs/base/release_notes.rst
@@ -29,6 +29,10 @@ Version 3.21.2
 - Update docs rest API
 - Improve log function (custom logger behavior / creator)
 - Fix superworkflow bug for standalone services
+- Dont display private properties in parameterized run results
+- Add Ansible playbook service log to security logger
+- Update superworkflow initial payload with placeholder service initial payload
+- Dont update netmiko and napalm configuration / oper data backup if empty result / no commands
 
 Version 3.21.1
 --------------